potential fix for peak_local_max

+++ src/napari_foobar/_widget.py
@@ -397,7 +397,7 @@ def widget_wrapper():
         labels_eroded=cv.erode(np.float32(label_isolated),kernel,iterations=1)
         other_labels=np.where(labels!=widget.labeled_cells[widget.current_index],labels,0)
         dist = ndi.distance_transform_edt(labels_eroded) #make distance map
-        coords = peak_local_max(dist,footprint=np.ones((50, 50,50)), labels=labels_eroded)#.astype('int64'))
+        coords = peak_local_max(dist,footprint=np.ones((50, 50,50)), labels=labels_eroded.astype('int32'))#.astype('int64'))
         mask = np.zeros(dist.shape, dtype=bool)
         mask[tuple(coords.T)] = True
         markers, _ = ndi.label(mask,structure=ndi.generate_binary_structure(3, 3))#3d-image(26) connectivity
